Spans as linked list, population of span values
- Spans are built into a linked list for easy navigation to the next one
when accessing their forward differenced values from the CubicBezier
parent objects.
- Setting an h forward differencing increment value sets Span h
increments such that they all have equal x intervals for even movement
through the curve in a functional context. Note--This requires that it
be possible for all Spans to be represented as functions of x. In the
context of x being a time variable in a motion path, that should never
happen anyway, for obvious reasons.